% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTSVtoXLSX.R
\name{convertTSVtoXLSX}
\alias{convertTSVtoXLSX}
\title{Convert TSV files to XLSX files}
\usage{
convertTSVtoXLSX()
}
\description{
This function converts every TSV file found in the \verb{5_analyzeROI > output} directory into their respective XLSX files.#'
}
