% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkspaceStructure.R
\name{createWorkspaceStructure}
\alias{createWorkspaceStructure}
\title{Setup the working directory}
\usage{
createWorkspaceStructure(wd = wd)
}
\arguments{
\item{wd}{The path in which the blank workspace will be created.}
}
\description{
This function setups the workspace structure. Within the specified \code{wd} path, every directory and subsequent content will be wiped and several empty directories and subdirectories will be created.
}
