% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSigmaPairsList.R
\name{generateSigmaPairsList}
\alias{generateSigmaPairsList}
\title{Generate a list of different sigma pairs to benchmark}
\usage{
generateSigmaPairsList(radius = NULL, maxSigmaRatio = NULL, step = 0.1)
}
\arguments{
\item{radius}{The estimated overall cell radius mean (in µm). This value should be measured under ImageJ.}

\item{maxSigmaRatio}{The maximum possible value of the \code{sigmaHigh/sigmaLow} ratio.}

\item{step}{The step to use to compute the sequence between 1 (excluded) and the \code{maxSigmaRatio} argument.}
}
\description{
This function generates a list of different sigma pairs to benchmark. The function uses the \code{radius} argument to estimate the base \code{sigmaLow} value to use (using the \verb{radius = (sigma * (sqrt(2*log(255, 2)))) - 1 relation)}, then generates different \code{sigmaHigh} values by multiplying \code{sigmaLow} with different multiplicators specified by the \code{maxSigmaRatio} and the \code{step} arguments. For instance, a \code{maxSigmaRatio} argument set to \code{2} and a \code{step} argument set to \code{0.25} will generate the following multiplicators: \code{c(1.25, 1.5, 1.75, 2)} which will be multiplied by the computed \code{sigmaLow} value, thus leading to 4 different sigma pairs to test.
}
