% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivePUPAID.R
\name{interactivePUPAID}
\alias{interactivePUPAID}
\title{Shiny app allowing the interactive use of PUPAID workflow}
\usage{
interactivePUPAID()
}
\description{
This Shiny app can be used instead of the classical script approach in order to use PUPAID. It will open a new window which will present the overall workflow of PUPAID. It is composed of buttons to automatically launch the necessary functions as well as text inputs to enter requested values. It also gives advice and tips about the analysis workflow and how to correctly use it. This application is ideal for beginners or for people who are not fluent with R programming or programming in general.
}
