% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTSVtoFCS.R
\name{mergeTSVtoFCS}
\alias{mergeTSVtoFCS}
\title{Merge TSV files to FCS}
\usage{
mergeTSVtoFCS()
}
\description{
This function merges every TSV file found in the \verb{5_analyzeROI > output} directory into a single FCS file. This FCS file could be directly used in standard flow cytometry analysis softwares such as FlowJo, Kaluza or others.
}
