% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameAndOrderTiles.R
\name{renameAndOrderTiles}
\alias{renameAndOrderTiles}
\title{Rename and order tiles for stitching}
\usage{
renameAndOrderTiles(AOI_x, AOI_y)
}
\arguments{
\item{AOI_x}{The width of a single tile (in real length unit, not in pixels).}

\item{AOI_y}{The height of a single tile (in real length unit, not in pixels).}
}
\value{
A list containing information about identified ROI (sample name, dimensions, tiles, etc.).
}
\description{
This function renames and orders every tile found in the \verb{2_renameAndOrderTiles > input} directory. The function will automatically determine the embedded ROI using the tiles coordinates and separate them accordingly in dedicated directories. It will also rename the tiles in a way suitable for ImageJ stitching.
}
