% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMacro_analyzeROI.R
\name{runMacro_analyzeROI}
\alias{runMacro_analyzeROI}
\title{Analyze ROI}
\usage{
runMacro_analyzeROI(
  imageJPath,
  wd,
  blockSize = 30,
  histogramBins = 256,
  maximumSlope = 3,
  sigmaLow,
  sigmaHigh
)
}
\arguments{
\item{imageJPath}{The absolute path where the ImageJ executable is located.}

\item{wd}{The workspace directory root path.}

\item{blockSize}{A numeric value which represents the block size to use for the CLAHE algorithm.}

\item{histogramBins}{A numeric value which represents the histogram bins to use for the CLAHE algorithm.}

\item{maximumSlope}{A numeric value which represents the maximum slope to use for the CLAHE algorithm.}

\item{sigmaLow}{A numeric value which represents the lowest sigma value to use for cell contouring.}

\item{sigmaHigh}{A numeric value which represents the highest sigma value to use for cell contouring.}
}
\description{
This function analyzes the ROI located in the \verb{5_analyzeROI > input} directory. First, the user will be asked to rename each channel embedded within the ROI according to a specific format (see below), and then select which channel to use for cell contouring. Then, data within each channel will be independently transformed to remove background noise and locally enhance contrast. Afterwards, Difference of Gaussians (DoG) method will be applied with the desired \code{sigmaLow} and \code{sigmaHigh} values for cell contouring. Then, the transformed signals for each remaining channel will be measured inside each determined cell before exportation in TSV format. In parallel, processed images will also be produced and saved. All the outputs will be located in the \verb{5_analyzeROI > output} directory. Please note that because of some technical limitations, this function will only launch a new instance of ImageJ without launching the actual analysis macro. Instead, the user should run \verb{Plugins > Macros > Run...} within ImageJ in order to launch the edited macro of interest named \verb{5_analyzeROI_run.txt} and located in the \code{macros} directory.
}
