% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMacro_generateBlackTile.R
\name{runMacro_generateBlackTile}
\alias{runMacro_generateBlackTile}
\title{Generate a black tile}
\usage{
runMacro_generateBlackTile(imageJPath, wd)
}
\arguments{
\item{imageJPath}{The absolute path where the ImageJ executable is located.}

\item{wd}{The workspace directory root path.}
}
\description{
This function generates a black tile devoid of any signal in each channel. The source tile should be placed in the \verb{1_generateBlackTile > input} directory, and the produced black tile will be created in the \verb{1_generateBlackTile > output} directory.
}
