% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMacro_stitchTiles.R
\name{runMacro_stitchTiles}
\alias{runMacro_stitchTiles}
\title{Stitch tiles from a single ROI}
\usage{
runMacro_stitchTiles(
  sampleName,
  currentROI_nb,
  currentROI_xSize,
  currentROI_ySize,
  imageJPath,
  wd
)
}
\arguments{
\item{sampleName}{The name of the treated sample. Should be obtained through the \code{stitchTiles()} function.}

\item{currentROI_nb}{The number of the current ROI to stitch. Should be obtained through the \code{stitchTiles()} function.}

\item{currentROI_xSize}{The number of tiles (X axis) for the current ROI. Should be obtained through the \code{stitchTiles()} function.}

\item{currentROI_ySize}{The number of tiles (Y axis) for the current ROI. Should be obtained through the \code{stitchTiles()} function.}

\item{imageJPath}{The absolute path where the ImageJ executable is located. Should be obtained through the \code{stitchTiles()} function.}

\item{wd}{The workspace directory root path. Should be obtained through the \code{stitchTiles()} function.}
}
\description{
This function stitches in ImageJ all the tiles found in the \verb{2_renameAndOrderTiles > output} directory (for a single given ROI). The stitched ROI will be generated in the \verb{3_stitchedTiles} directory. This function should only be launched through the \code{stitchTiles()} function.
}
