% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMacro_testSigmaPair.R
\name{runMacro_testSigmaPair}
\alias{runMacro_testSigmaPair}
\title{Test a given sigma pair}
\usage{
runMacro_testSigmaPair(wd, imageJPath, sigmaLow, sigmaHigh)
}
\arguments{
\item{wd}{The workspace directory root path.}

\item{imageJPath}{The absolute path where the ImageJ executable is located.}

\item{sigmaLow}{A numeric value which represents the lowest sigma value to use for cell contouring.}

\item{sigmaHigh}{A numeric value which represents the highest sigma value to use for cell contouring.}
}
\description{
This function performs cell contouring (using Different of Gaussians, (DoG) method) in ImageJ with a given pair of sigma values. This function expects an image file within the \verb{4_testSigmaPair > input} directory with only 1 channel that you should first generate manually using ImageJ, for instance. The selected signal should be viable for automatized cell contouring, like any nucleus-restricted stain (such as DAPI) or any cell membrane-restricted stain. The selected image should be typically representative of an area with a high density of cells, in order to better estimate if the used pair of sigma values allows to clearly delineate the cells contours. For faster computation, the image could be a cropped region of the whole ROI of interest.
}
