% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchTiles.R
\name{stitchTiles}
\alias{stitchTiles}
\title{Stitch tiles}
\usage{
stitchTiles(
  renameAndOrderOutput = renameAndOrderOutput,
  imageJPath = imageJPath,
  wd = wd
)
}
\arguments{
\item{renameAndOrderOutput}{Tiles renaming and ordering output produced by \code{renameAndOrderTiles()} function.}

\item{imageJPath}{The absolute path where the ImageJ executable is located.}

\item{wd}{The workspace directory root path.}
}
\description{
This function stitches in ImageJ all the tiles found in the \verb{2_renameAndOrderTiles > output} directory (each ROI separately). The stitched ROI will be generated in the \verb{3_stitchedTiles} directory, as well as a single text file per ROI containing the tile configuration used during stitching. The function also outputs a rds file containing the determined ROI and their characteristics (height, width, number of total tiles, number of missing tiles, etc.).
}
